16-Polling System Smart Contract /polling-smart-contract/requirements.txt
web3==6.0.0
python-dotenv==1.0.0
py-solc-x==2.0.1